Had to go back before the massive clean up, but now the program gives correct results

+++ src/Basis/basis.cpp
@@ -7,9 +7,9 @@ long long Basis::factorial(const int n) {
     return (n == 1 || n == 0) ? 1 : factorial(n - 1) * n;
 }
 
-double Basis::binomial(const int n, const int p) {
+int Basis::binomial(const int n, const int p) {
     //Binomial coefficients, equal to magic numbers
-    return factorial(n+p)/(factorial(n)*factorial(p));
+    return int(factorial(n+p)/(factorial(n)*factorial(p)));
 }
 
 Basis::~Basis() {};
